Remove protectable from contributions edit+write
The confirmation dialog is opened on our staging environment when you try to
create a contribution. A successful update or create will redirect you to the
contribution show page, thus trigger beforeRouteLeave and opening the
confirmation dialog. The greatest value of the confirmation dialog is currently
on the comments, so I will remove it from the contribution for now.