Follow @appinteractive review
```
session on the server side (SSR) is not destroyed which leads to the following
scenario:

1. login as admin (test@test.de)
2. logout
3. login as regular user (test3@test3.de)
4. refresh page after you are logged in
5. see that you are logged in again as the last user (admin) instead of your
   last loggin (user)
You need to remove that session also on the server side, maybe there is a cookie
still flying around in the browser which is then send on SSR and the backend
responds with the logged in user from the token in the cookie.
```

This is not possible now, as the lazy-loading happens with `app.$api` and
not the socket.